Added support for Widened Compass
The textbox now takes widened compass into account when displaying
distances for out of range bubbles.

For indicolures, this buff is encoded in the indi effect for the
pc update and player update packets using the 0x20 bit of the field.

For geocolures, this is encoded in npc update packets.
If _unknown2 is 33536 (0x8300) it is a normal sized luopan, or
if _unknown2 is 34560 (0x8700) then it is a widened compass luopan.
One might expect to be able to check model_size or model_scale in
the mob table for this information, but it does not seem to be
reflected there.

Since listening for npc update packets (0x00E) would be annoying,
the code takes the less reliable approach of listening for the
widened compass action packet instead (as it does for other JAs).